+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - ILIAS 5.4 support
 - Drop ILIAS 5.1 and ILIAS 5.2 support
 - Implement uninstaller
+- Deliver for the SrUserEnrolment plugin a metadata field operator rule
 - Fix cancel object mapping button
 - Fix no array pass to `array_values`
 - Fix info block only show on info screen